requirements.txt
bcrypt==4.0.0
falcon==3.0.0
falcon-cors==1.1.7
falcon-multipart==0.2.0
falcon-swagger-ui==1.2.2
gunicorn==20.1.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.0.3
pkg_resources==0.0.0
psutil==5.9.2
PyJWT==1.5.3
pymongo==4.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
semantic-version==2.10.0
setuptools-rust==1.1.2
SQLAlchemy==1.3.23
typing_extensions==4.1.1
